requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-multipart==0.0.20
faiss-cpu==1.11.0.post1
numpy==2.3.4
openai==1.107.0
python-dotenv==1.1.1
pytest==8.4.2
pytest-cov==7.0.0
httpx==0.28.1
